+++ CHANGELOG.md
@@ -8,4 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Add a LMSHandler class to select the right LMS Backend to use according to
+  the course run's `resource_link` provided
+- Add a OpenEdX LMS Backend to manage enrollments
+
 [unreleased]: https://github.com/openfun/joanie